beginning race deltas
