+++ agg_funcs.py
@@ -2,6 +2,9 @@
 import os
 import psycopg2
 import pandas as pd
+
+race_list = {'1': 'native', '2': 'asian', '3': 'black', '4': 'hawaiian', '5': 'white', '6': 'not_provided', '7': 'not_applicable', '8': 'no_co_app'}
+race_dict = {'1': 'native', '2': 'asian', '3': 'black', '4': 'hawaiian', '5': 'white'}
 def agg_SQL(source_table, action):
 	#FIXME change action to a passed format variable
 	"""returns SQL_base with source table formatted into the query text
@@ -130,3 +133,118 @@ def agg_df(source_table, action, conn):
 def check_path(path):
 	if not os.path.exists(path):
 		os.makedirs(path)
+
+def format_load_SQL(table, data):
+	SQL = """COPY {table} FROM '{path}' DELIMITER ',' CSV HEADER; COMMIT;"""
+	return SQL.format(table=table, path=data)
+
+def create_aggregate_table_SQL(table, action):
+	SQL = """CREATE TABLE {create_table} (
+		year real,
+		state varchar(2),
+		county varchar(5),
+		fips varchar(5),
+		loan_average_{action} real,
+		income_average_{action} real,
+		count_{action} real,
+		value_{action} real,
+		native_loan_average_{action} real,
+		native_income_average_{action} real,
+		native_count_{action} real,
+		native_value_{action} real,
+		black_loan_average_{action} real,
+		black_income_average_{action} real,
+		black_count_{action} real,
+		black_value_{action} real,
+		asian_loan_average_{action} real,
+		asian_income_average_{action} real,
+		asian_count_{action} real,
+		asian_value_{action} real,
+		white_loan_average_{action} real,
+		white_income_average_{action} real,
+		white_count_{action} real,
+		white_value_{action} real,
+		hawaiian_loan_average_{action} real,
+		hawaiian_income_average_{action} real,
+		hawaiian_count_{action} real,
+		hawaiian_value_{action} real,
+		not_applicable_loan_average_{action} real,
+		not_applicable_income_average_{action} real,
+		not_applicable_count_{action} real,
+		not_applicable_value_{action} real,
+		not_provided_loan_average_{action} real,
+		not_provided_income_average_{action} real,
+		not_provided_count_{action} real,
+		not_provided_value_{action} real,
+		no_co_app_loan_average_{action} real,
+		no_co_app_income_average_{action} real,
+		no_co_app_count_{action} real,
+		no_co_app_value_{action} real);
+		COMMIT;"""
+	return SQL.format(create_table=table, action=action)
+
+def county_years_SQL(app_table, orig_table, fips):
+	SQL = """SELECT
+		app.year
+		,app.state
+		,app.county
+		,app.fips
+		,loan_average_orig
+		,income_average_orig
+		,count_orig
+		,value_orig
+		,native_loan_average_orig
+		,native_income_average_orig
+		,native_count_orig
+		,native_value_orig
+		,black_loan_average_orig
+		,black_income_average_orig
+		,black_count_orig
+		,black_value_orig
+		,asian_loan_average_orig
+		,asian_income_average_orig
+		,asian_count_orig
+		,asian_value_orig
+		,white_loan_average_orig
+		,white_income_average_orig
+		,white_count_orig
+		,white_value_orig
+		,hawaiian_loan_average_orig
+		,hawaiian_income_average_orig
+		,hawaiian_count_orig
+		,hawaiian_value_orig
+		,loan_average_app
+		,income_average_app
+		,count_app
+		,value_app
+		,native_loan_average_app
+		,native_income_average_app
+		,native_count_app
+		,native_value_app
+		,black_loan_average_app
+		,black_income_average_app
+		,black_count_app
+		,black_value_app
+		,asian_loan_average_app
+		,asian_income_average_app
+		,asian_count_app
+		,asian_value_app
+		,white_loan_average_app
+		,white_income_average_app
+		,white_count_app
+		,white_value_app
+		,hawaiian_loan_average_app
+		,hawaiian_income_average_app
+		,hawaiian_count_app
+		,hawaiian_value_app
+		FROM {app_table} AS app
+		FULL OUTER JOIN {orig_table} AS orig
+		ON app.fips = orig.fips
+		WHERE app.fips = cast({fips} AS VARCHAR(5))"""
+	return SQL.format(app_table=app_table, orig_table=orig_table, fips=fips)
+
+def get_CBSA_df(CBSA_file, seperator):
+	cbsa_df = pd.read_csv(CBSA_file, sep=seperator)
+	cbsa_df.county = cbsa_df.county.map(lambda x: str(x).zfill(5)) #left pad with 0's to make valid FIPS codes
+	return cbsa_df
+
